Removed default 100 Limit for Vertex URL task Fetch and code review changes
