+++ app/com/linkedin/drelephant/tez/fetchers/TezFetcher.java
@@ -17,27 +17,21 @@ package com.linkedin.drelephant.tez.fetchers;
 import com.linkedin.drelephant.analysis.AnalyticJob;
 import com.linkedin.drelephant.analysis.ElephantFetcher;
 import com.linkedin.drelephant.configurations.fetcher.FetcherConfigurationData;
-import com.linkedin.drelephant.math.Statistics;
 import com.linkedin.drelephant.tez.data.TezApplicationData;
 import com.linkedin.drelephant.tez.data.TezCounterData;
 import com.linkedin.drelephant.tez.data.TezTaskData;
+import com.linkedin.drelephant.util.ThreadContextMR2;
 import org.apache.log4j.Logger;
 
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.security.authentication.client.AuthenticatedURL;
 import org.apache.hadoop.security.authentication.client.AuthenticationException;
 import org.codehaus.jackson.JsonNode;
-import org.codehaus.jackson.map.ObjectMapper;
 
 import java.io.IOException;
-import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.net.URLConnection;
 import java.util.*;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 /**
  * Task level data mining for Tez Tasks from timeline server API
@@ -46,7 +40,7 @@ import java.util.regex.Pattern;
 public class TezFetcher implements ElephantFetcher<TezApplicationData> {
 
   private static final Logger logger = Logger.getLogger(TezFetcher.class);
-  private static final int MAX_SAMPLE_SIZE = 1000;
+
   private static final String TIMELINE_SERVER_URL = "yarn.timeline-service.webapp.address";
 
   private URLFactory _urlFactory;
@@ -109,16 +103,14 @@ public class TezFetcher implements ElephantFetcher<TezApplicationData> {
             reducerListAggregate = reducerList;
             maxSize = mapperList.size() + reducerList.size();
           }
-
+        }
+        if (state.equals("FAILED")) {
+          jobData.setSucceeded(false);
         }
       }
       finally {
         ThreadContextMR2.updateAuthToken();
       }
-    }
-
-    if(mapperListAggregate.isEmpty() && reducerListAggregate.isEmpty()){
-      jobData.setSucceeded(false);
     }
 
     TezTaskData[] mapperData = mapperListAggregate.toArray(new TezTaskData[mapperListAggregate.size()]);
@@ -177,7 +169,7 @@ public class TezFetcher implements ElephantFetcher<TezApplicationData> {
 
     private URL getTaskListByVertexURL(String dagId, String vertexId) throws MalformedURLException {
       return new URL(_timelineWebAddr + "/TEZ_TASK_ID?primaryFilter=TEZ_DAG_ID:" + dagId +
-          "&secondaryFilter=TEZ_VERTEX_ID:" + vertexId);
+          "&secondaryFilter=TEZ_VERTEX_ID:" + vertexId + "&limit=500000");
     }
 
     private URL getTasksURL(String taskId) throws MalformedURLException {
@@ -336,13 +328,7 @@ public class TezFetcher implements ElephantFetcher<TezApplicationData> {
     private void getTaskData(String dagId, List<TezTaskData> taskList, boolean isMapTask)
         throws IOException, AuthenticationException {
 
-      if (taskList.size() > MAX_SAMPLE_SIZE) {
-        logger.info(dagId + " needs sampling.");
-        Collections.shuffle(taskList);
-      }
-
-      int sampleSize = Math.min(taskList.size(), MAX_SAMPLE_SIZE);
-      for (int i=0; i<sampleSize; i++) {
+      for(int i=0; i<taskList.size(); i++) {
         TezTaskData data = taskList.get(i);
         URL taskCounterURL = getTaskURL(data.getTaskId());
         TezCounterData taskCounter = getTaskCounter(taskCounterURL);
@@ -400,82 +386,6 @@ public class TezFetcher implements ElephantFetcher<TezApplicationData> {
       long[] time = new long[] { finishTime - startTime, shuffleTime, mergeTime, startTime, finishTime };
 
       return time;
-    }
-  }
-}
-
-final class ThreadContextMR2 {
-  private static final Logger logger = Logger.getLogger(ThreadContextMR2.class);
-
-  private static final AtomicInteger THREAD_ID = new AtomicInteger(1);
-
-  private static final ThreadLocal<Integer> _LOCAL_THREAD_ID = new ThreadLocal<Integer>() {
-    @Override
-    public Integer initialValue() {
-      return THREAD_ID.getAndIncrement();
-    }
-  };
-
-  private static final ThreadLocal<Long> _LOCAL_LAST_UPDATED = new ThreadLocal<Long>();
-  private static final ThreadLocal<Long> _LOCAL_UPDATE_INTERVAL = new ThreadLocal<Long>();
-
-  private static final ThreadLocal<Pattern> _LOCAL_DIAGNOSTIC_PATTERN = new ThreadLocal<Pattern>() {
-    @Override
-    public Pattern initialValue() {
-      // Example: "Task task_1443068695259_9143_m_000475 failed 1 times"
-      return Pattern.compile(
-          "Task[\\s\\u00A0]+(.*)[\\s\\u00A0]+failed[\\s\\u00A0]+([0-9])[\\s\\u00A0]+times[\\s\\u00A0]+");
-    }
-  };
-
-  private static final ThreadLocal<AuthenticatedURL.Token> _LOCAL_AUTH_TOKEN =
-      new ThreadLocal<AuthenticatedURL.Token>() {
-        @Override
-        public AuthenticatedURL.Token initialValue() {
-          _LOCAL_LAST_UPDATED.set(System.currentTimeMillis());
-          // Random an interval for each executor to avoid update token at the same time
-          _LOCAL_UPDATE_INTERVAL.set(Statistics.MINUTE_IN_MS * 30 + new Random().nextLong()
-              % (3 * Statistics.MINUTE_IN_MS));
-          logger.info("Executor " + _LOCAL_THREAD_ID.get() + " update interval " + _LOCAL_UPDATE_INTERVAL.get() * 1.0
-              / Statistics.MINUTE_IN_MS);
-          return new AuthenticatedURL.Token();
-        }
-      };
-
-  private static final ThreadLocal<AuthenticatedURL> _LOCAL_AUTH_URL = new ThreadLocal<AuthenticatedURL>() {
-    @Override
-    public AuthenticatedURL initialValue() {
-      return new AuthenticatedURL();
-    }
-  };
-
-  private static final ThreadLocal<ObjectMapper> _LOCAL_MAPPER = new ThreadLocal<ObjectMapper>() {
-    @Override
-    public ObjectMapper initialValue() {
-      return new ObjectMapper();
-    }
-  };
-
-  private ThreadContextMR2() {
-    // Empty on purpose
-  }
-
-  public static Matcher getDiagnosticMatcher(String diagnosticInfo) {
-    return _LOCAL_DIAGNOSTIC_PATTERN.get().matcher(diagnosticInfo);
-  }
-
-  public static JsonNode readJsonNode(URL url) throws IOException, AuthenticationException {
-    HttpURLConnection conn = _LOCAL_AUTH_URL.get().openConnection(url, _LOCAL_AUTH_TOKEN.get());
-    return _LOCAL_MAPPER.get().readTree(conn.getInputStream());
-  }
-
-  public static void updateAuthToken() {
-    long curTime = System.currentTimeMillis();
-    if (curTime - _LOCAL_LAST_UPDATED.get() > _LOCAL_UPDATE_INTERVAL.get()) {
-      logger.info("Executor " + _LOCAL_THREAD_ID.get() + " updates its AuthenticatedToken.");
-      _LOCAL_AUTH_TOKEN.set(new AuthenticatedURL.Token());
-      _LOCAL_AUTH_URL.set(new AuthenticatedURL());
-      _LOCAL_LAST_UPDATED.set(curTime);
     }
   }
 }
