+++ test/com/linkedin/drelephant/tez/fetchers/TezFetcherTest.java
@@ -19,15 +19,17 @@ package com.linkedin.drelephant.tez.fetchers;
 import java.util.regex.Matcher;
 import org.junit.Assert;
 import org.junit.Test;
+import com.linkedin.drelephant.util.ThreadContextMR2;
 
 public class TezFetcherTest {
 
   @Test
   public void testDiagnosticMatcher() {
-    Matcher matcher = com.linkedin.drelephant.tez.fetchers.ThreadContextMR2.getDiagnosticMatcher("Task task_1443068695259_9143_m_000475 failed 1 time");
-    Assert.assertEquals("Task[\\s\\u00A0]+(.*)[\\s\\u00A0]+failed[\\s\\u00A0]+([0-9])[\\s\\u00A0]+times[\\s\\u00A0]+", matcher.pattern().toString());
-    Assert.assertEquals(false, matcher.matches());
-    Assert.assertEquals(2, matcher.groupCount());
+    Matcher matcher = ThreadContextMR2.getDiagnosticMatcher("Task task_1443068695259_9143_m_000475 failed 1 time");
+    Assert.assertEquals(".*[\\s\\u00A0]+(task_[0-9]+_[0-9]+_[m|r]_[0-9]+)[\\s\\u00A0]+.*", matcher.pattern().toString());
+    Assert.assertEquals(true, matcher.matches());
+    Assert.assertEquals(1, matcher.groupCount());
+    Assert.assertEquals("task_1443068695259_9143_m_000475", matcher.group(1));
   }
 
 }
